+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/components/ConfigTaskPanel/AdvancedOptions.tsx
@@ -1,0 +1,70 @@
+import { IDropdownOption, Stack } from '@fluentui/react';
+import React from 'react';
+import { useSelector } from 'react-redux';
+import { State } from 'RootStateType';
+import { InferenceSource, ProjectData } from '../../store/project/projectTypes';
+import { ExpandPanel } from '../ExpandPanel';
+import { CameraFPSOptions } from './CameraFPSOptions';
+import { CloudMsgOption } from './CloudMsgOption';
+import { ProtocolOptions } from './ProtocolOptions';
+import { RetrainImgOption } from './RetrainImgOption';
+import { SendVideoOptions } from './SendVideoOptions';
+import { OnChangeType } from './type';
+
+type AdvancedOptionsProps = {
+  projectData: ProjectData;
+  selectedCameraOptions: IDropdownOption[];
+  selectedPartOptions: IDropdownOption[];
+  onChange: OnChangeType;
+};
+
+export const AdvancedOptions: React.FC<AdvancedOptionsProps> = ({
+  projectData,
+  selectedCameraOptions,
+  selectedPartOptions,
+  onChange,
+}) => {
+  const selectedTrainProjectIsNotDemo = useSelector((state: State) =>
+    state.trainingProject.nonDemo.includes(projectData.trainingProject),
+  );
+  return (
+    <ExpandPanel titleHidden="Advanced options" titleVisible="Fewer options" iconPosition="end" bottomBorder>
+      <Stack horizontal tokens={{ childrenGap: 50 }} wrap styles={{ root: { paddingTop: '30px' } }}>
+        <CloudMsgOption
+          sendMessageToCloud={projectData.sendMessageToCloud}
+          framesPerMin={projectData.framesPerMin}
+          onChange={onChange}
+        />
+        {selectedTrainProjectIsNotDemo && (
+          <RetrainImgOption
+            needRetraining={projectData.needRetraining}
+            accuracyRangeMin={projectData.accuracyRangeMin}
+            accuracyRangeMax={projectData.accuracyRangeMax}
+            maxImages={projectData.maxImages}
+            onChange={onChange}
+          />
+        )}
+        {projectData.inferenceSource === InferenceSource.LVA && (
+          <>
+            <CameraFPSOptions
+              setFpsManually={projectData.setFpsManually}
+              fps={projectData.fps}
+              recomendedFps={projectData.recomendedFps}
+              onChange={onChange}
+            />
+            <SendVideoOptions
+              SVTCisOpen={projectData.SVTCisOpen}
+              SVTCcameras={projectData.SVTCcameras}
+              SVTCparts={projectData.SVTCparts}
+              SVTCconfirmationThreshold={projectData.SVTCconfirmationThreshold}
+              onChange={onChange}
+              selectedCameraOptions={selectedCameraOptions}
+              selectedPartOptions={selectedPartOptions}
+            />
+          </>
+        )}
+        <ProtocolOptions inferenceProtocol={projectData.inferenceProtocol} onChange={onChange} />
+      </Stack>
+    </ExpandPanel>
+  );
+};
