+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/components/ConfigTaskPanel/type.ts
@@ -1,0 +1,3 @@
+import { ProjectData } from '../../store/project/projectTypes';
+
+export type OnChangeType = <K extends keyof P, P = ProjectData>(key: K, value: P[K]) => void;
